Bug fixes, etc.
If multiple servers reported the exact same lat. and long. then they
would overwrite eachother in the dict.  I changed it to use a
defaultdict so that you could have multiple servers with the exact same
distance. (3 of my closest 5 reported the exact same corrdinates.)

The latency it was printing was not the latency from the chosen best
server.  Now it is.

Added a print out of the IP and ISP info.