Added function for simple module-like usage
The speedtest_cli code is awesome, but not written to be consumed as a
module. To facilitate its usage by other developers, a simple function -
basic_speedtest() - has been added that returns the speedtest data as a
simple tuple. This change is additive and restricted to default/silent
options. A proper refactoring could be done to enable such module-like
usage the correct way, but this is an acceptable, safe, and
easy-to-validate alternative.