Correct average latency calculation in best server determination.
Reverts part of commit c1b9a0d. I guess some exploratory or debugging edits
made it into the commit.

Three samples are run, summed, and divided by six. It should be three.

As a result, the latency results were halved. The "best" server was still
chosen but the ping time displayed was half of the real average result.